[Canvas] Moves canvas Server to NP completely (#65148)
* Moves canvas Server to NP completely

* Fix typecheck
