+++ x-pack/legacy/plugins/canvas/canvas_plugin_src/functions/server/escount.ts
@@ -8,8 +8,10 @@ import {
   ExpressionFunctionDefinition,
   ExpressionValueFilter,
 } from 'src/plugins/expressions/common';
+/* eslint-disable */
 // @ts-ignore untyped local
-import { buildESRequest } from '../../../server/lib/build_es_request';
+import { buildESRequest } from '../../../../../../plugins/canvas/server/lib/build_es_request';
+/* eslint-enable */
 import { getFunctionHelp } from '../../../i18n';
 
 interface Arguments {
