+++ x-pack/legacy/plugins/canvas/canvas_plugin_src/functions/server/esdocs.ts
@@ -6,8 +6,10 @@
 
 import squel from 'squel';
 import { ExpressionFunctionDefinition } from 'src/plugins/expressions';
+/* eslint-disable */
 // @ts-ignore untyped local
-import { queryEsSQL } from '../../../server/lib/query_es_sql';
+import { queryEsSQL } from '../../../../../../plugins/canvas/server/lib/query_es_sql';
+/* eslint-enable */
 import { ExpressionValueFilter } from '../../../types';
 import { getFunctionHelp } from '../../../i18n';
