+++ x-pack/legacy/plugins/canvas/canvas_plugin_src/functions/server/essql.ts
@@ -5,8 +5,10 @@
  */
 
 import { ExpressionFunctionDefinition } from 'src/plugins/expressions/common';
+/* eslint-disable */
 // @ts-ignore untyped local
-import { queryEsSQL } from '../../../server/lib/query_es_sql';
+import { queryEsSQL } from '../../../../../../plugins/canvas/server/lib/query_es_sql';
+/* eslint-enable */
 import { ExpressionValueFilter } from '../../../types';
 import { getFunctionHelp } from '../../../i18n';
