+++ x-pack/legacy/plugins/canvas/index.js
@@ -6,7 +6,6 @@
 
 import { resolve } from 'path';
 import { DEFAULT_APP_CATEGORIES } from '../../../../src/core/utils';
-import { init } from './init';
 import { CANVAS_APP, CANVAS_TYPE, CUSTOM_ELEMENT_TYPE } from './common/lib';
 
 export function canvas(kibana) {
@@ -64,6 +63,6 @@ export function canvas(kibana) {
       }).default();
     },
 
-    init,
+    init: () => undefined,
   });
 }
