+++ x-pack/plugins/canvas/server/lib/build_bool_array.js
