+++ x-pack/plugins/canvas/server/lib/filters.js
