+++ x-pack/plugins/canvas/server/lib/format_response.js
