+++ x-pack/plugins/canvas/server/lib/get_es_filter.js
