+++ x-pack/plugins/canvas/server/lib/normalize_type.js
