+++ x-pack/plugins/canvas/server/lib/sanitize_name.js
